hotfix: 답글 삭제 버그 수정, 답글열기 버그 수정
